pyproject.toml
[project]
name = "processpype"
version = "2.0.0"
description = "A modular process environment framework: observability, notifications, server, and services — all driven by YAML config"
authors = [{ name = "Gianluca Pagliara", email = "pagliara.gianluca@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.10.4",
    "pyyaml>=6.0.2",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
events = ["eventspype>=1.1.0,<2"]
tracing = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]
logfire = [
    "logfire>=4.16",
    "processpype[tracing]",
]
otlp = [
    "opentelemetry-exporter-otlp>=1.20",
    "processpype[tracing]",
]
telegram = ["telethon>=1.35"]
full = [
    "processpype[events,logfire,telegram]",
]

[project.urls]
Repository = "https://github.com/gianlucapagliara/processpype"
Documentation = "https://gianlucapagliara.github.io/processpype/"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "pytest-timeout>=2.3",
    "mypy>=1.7.0",
    "ruff>=0.8.4",
    "pre-commit>=4.0",
    "types-pyyaml>=6.0.12.20241230",
]
docs = [
    "mkdocs>=1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
strict = true
disallow_untyped_decorators = false
ignore_missing_imports = true
disable_error_code = ["misc"]
exclude = ["tests/.*"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
timeout = 30
testpaths = ["tests"]
